requirements.txt
langchain
streamlit
python-dotenv
google-generativeai
langchain-google-genai